DS-75: Cleaning part 1
