--- ngx-fudis/projects/ngx-fudis/src/lib/components/button/public-api.ts
@@ -1,1 +1,0 @@
-export * from './button.component';
