+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/checkbox/checkbox.component.spec.ts
@@ -21,10 +21,4 @@ describe('CheckboxComponent', () => {
 	it('should create', () => {
 		expect(component).toBeTruthy();
 	});
-
-	describe('checkbox has states', () => {
-		it('checking the check state of componen', () => {
-			component.checked = true;
-		});
-	});
 });
