+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/error-summary/error-summary.stories.ts
@@ -1,10 +1,7 @@
 import { Story, Meta } from '@storybook/angular/types-6-0';
 import { moduleMetadata } from '@storybook/angular';
-import { MatInputModule } from '@angular/material/input';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { VanillaTextInputComponent } from '../vanilla-text-input/vanilla-text-input.component';
+import { TextInputComponent } from '../text-input/text-input.component';
 import { HeadingComponent } from '../../typography/heading/heading.component';
 import { ButtonComponent } from '../../button/button.component';
 import { ErrorSummaryComponent } from './error-summary.component';
@@ -14,8 +11,8 @@ export default {
 	component: ErrorSummaryComponent,
 	decorators: [
 		moduleMetadata({
-			imports: [BrowserAnimationsModule, MatInputModule, MatFormFieldModule, ReactiveFormsModule, FormsModule],
-			declarations: [VanillaTextInputComponent, HeadingComponent, ButtonComponent],
+			imports: [ReactiveFormsModule, FormsModule],
+			declarations: [TextInputComponent, HeadingComponent, ButtonComponent],
 		}),
 	],
 	argTypes: {},
@@ -31,29 +28,29 @@ export const ErrorSummary: Story = () => ({
 	template: html`
 		<form id="testform">
 			<fudis-error-summary>
-				<fudis-vanilla-text-input
+				<fudis-text-input
 					[required]="true"
 					labelHeight="double"
 					label="Password, min length 4, max 6"
 					minLength="4"
 					maxLength="6"
 					type="password"
-					helpText="Help text Nipperkin yardarm list splice the main brace draft swing the lead Privateer tack pillage bounty. Shrouds quarter gangplank belaying pin skysail fathom rope's end boom pirate spike. Topgallant line barque sloop quarter cog Jack Tar gabion pillage fire ship."></fudis-vanilla-text-input>
-				<fudis-vanilla-text-input
+					helpText="Help text Nipperkin yardarm list splice the main brace draft swing the lead Privateer tack pillage bounty. Shrouds quarter gangplank belaying pin skysail fathom rope's end boom pirate spike. Topgallant line barque sloop quarter cog Jack Tar gabion pillage fire ship."></fudis-text-input>
+				<fudis-text-input
 					[required]="true"
 					labelHeight="double"
 					label="Label Email, pakollinen, Nipperkin yardarm list splice the main brace draft."
 					type="email"
-					helpText="Lorem ipsum"></fudis-vanilla-text-input>
-				<fudis-vanilla-text-input
+					helpText="Lorem ipsum"></fudis-text-input>
+				<fudis-text-input
 					label="Email, ei-pakollinen, mutta olen pitkä kuin suomen talvi ja valoisaa aikaa on noin 13 sekuntia"
 					type="email"
-					helpText="Lorem ipsum"></fudis-vanilla-text-input>
-				<fudis-vanilla-text-input
+					helpText="Lorem ipsum"></fudis-text-input>
+				<fudis-text-input
 					label="Tavallinen teksti, min length 4"
 					type="text"
 					minLength="5"
-					helpText="Lorem ipsum"></fudis-vanilla-text-input>
+					helpText="Lorem ipsum"></fudis-text-input>
 			</fudis-error-summary>
 		</form>
 	`,
