--- ngx-fudis/projects/ngx-fudis/src/lib/components/form/fudis-checkbox/fudis-checkbox.component.html
@@ -1,32 +1,0 @@
-<div class="fudis-checkbox" (click)="toggle()">
-	<label class="fudis-checkbox__label" for="{{ id }}">
-		<div class="fudis-checkbox__label__text">{{ label }}</div>
-		<div class="fudis-checkbox__box fudis-checkbox__box--{{ invalid }} fudis-checkbox__box--{{ disabledInput }}">
-			<input
-				[attr.aria-label]="label"
-				type="checkbox"
-				name="checkbox"
-				id="{{ id }}"
-				class="fudis-checkbox__input"
-				(blur)="checkErrors()"
-				(change)="checkErrors()"
-				[disabled]="disabled"
-				[required]="required" />
-			<svg
-				width="16"
-				height="16"
-				preserveAspectRatio="xMidYMin"
-				xmlns="http://www.w3.org/2000/svg"
-				viewBox="0 0 22 22"
-				id="check-small"
-				class="fudis-checkbox__icon"
-				aria-hidden="true"
-				*ngIf="checked">
-				<path
-					class="a"
-					d="M9.0171,16.5A1.5,1.5,0,0,1,7.9,16.0012l-4.0171-4.48a1.5,1.5,0,0,1,2.2334-2.003L8.8706,12.59l5.9-8.4487A1.5,1.5,0,1,1,17.23,5.8591l-6.9824,10a1.4978,1.4978,0,0,1-1.1392.6387C9.0776,16.4992,9.0474,16.5,9.0171,16.5Z" />
-			</svg>
-		</div>
-	</label>
-</div>
-<div *ngIf="showError" class="fudis-checkbox__guidance__error">Arvo on pakollinen</div>
