--- ngx-fudis/projects/ngx-fudis/src/lib/components/form/fudis-checkbox/fudis-checkbox.component.scss
@@ -1,80 +1,0 @@
-@use '../../../foundations/borders/mixins.scss' as borders;
-@use '../../../foundations/spacing/tokens.scss' as spacing;
-@use '../../../foundations/colors/tokens.scss' as colors;
-@use '../../../foundations/colors/mixins.scss' as colorMixins;
-@use '../../../foundations/focus/mixins.scss' as focus; 
-@use '../../../foundations/typography/mixins.scss' as typography;
-
-.fudis-checkbox {
-	display: inline-block;
-	padding-top: spacing.$fudis-spacing-xs;
-	width: auto;
-
-	&__box {
-		@include borders.fudis-border('1px', 'solid', 'gray-middle');
-
-		position: absolute;
-		margin: spacing.$fudis-spacing-xs;
-		border-radius: 0;
-		cursor: pointer;
-		width: spacing.$fudis-spacing-s;
-		height: spacing.$fudis-spacing-s;
-
-		:focus {
-			@include focus.fudis-focus-checkbox;
-		}
-
-		&--disabled {
-			@include borders.fudis-border('1px', 'dashed', 'gray-middle');	
-			@include colorMixins.fudis-bg-color('gray-extra-light');
-
-			cursor: default;
-		}
-
-		&--invalid {
-			@include borders.fudis-border('1px', 'solid', 'red');
-			
-			box-shadow: 0 0 0 spacing.$fudis-pixel-3 colors.$fudis-red-light;
-		}
-
-	}
-
-	&__input {
-		appearance: none;
-		display: grid;
-		place-content: center;
-		margin: 0;
-	}
-
-	&__icon {
-		position: absolute;
-		top: spacing.$fudis-pixel-1;
-		fill: colors.$fudis-gray-middle;
-	}
-
-
-	&__label {
-		@include typography.fudis-body-m-regular;
-		@include colorMixins.fudis-text-color('gray-dark');
-
-		display: flex;
-		align-items: center;
-		margin-bottom: 0;
-		padding-bottom: spacing.$fudis-spacing-xs;
-
-		&__text {
-			position: relative;
-			margin-left: spacing.$fudis-spacing-l;
-		}
-	}
-
-	&__guidance {
-		&__error {
-			@include typography.fudis-body-s-regular;
-			@include colorMixins.fudis-text-color('red');
-
-			margin-top: spacing.$fudis-spacing-xs;
-			margin-left: spacing.$fudis-spacing-xs;
-		}
-	}
-}
