+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/text-area/text-area.component.scss
