+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/text-input/text-input.component.html
@@ -1,57 +1,32 @@
-<mat-form-field class="fudis-text-input" floatLabel="always" appearance="standard">
-	<mat-label class="fudis-text-input__label">
-		{{ label }}
-	</mat-label>
-	<input
-		(blur)="checkErrors()"
-		(change)="checkErrors()"
-		#fudisTextInput
-		matInput
-		[type]="type"
-		[required]="required"
-		[formControl]="fudisFormControl"
-		[ngClass]="classes"
-		class="fudis-text-input__input"
-		[attr.minlength]="minLength"
-		[attr.maxlength]="maxLength" />
-	<mat-error class="fudis-text-input__error" *ngIf="fudisFormControl.invalid && fudisFormControl.touched">
-		{{ this.defaultError }}
-	</mat-error>
-	<!-- 
-	<mat-error class="fudis-text-input__error" *ngIf="fudisFormControl.errors?.['minlength'] && fudisFormControl.touched">
-		Input must be at least {{ minLength }} long
-	</mat-error>
+<div class="fudis-form-element fudis-text-input fudis-text-input__{{ size }}">
+	<label class="fudis-text-input__label" for="{{ id }}">
+		<div #fudisTextInputLabel class="fudis-text-input__label__content">
+			<span class="fudis-text-input__label__text fudis-text-input__label__text__{{ labelHeight }}">{{ label }}</span
+			><span class="fudis-text-input__label__required" *ngIf="required">Pakollinen&nbsp;*</span>
+		</div>
+		<input
+			[attr.aria-describedby]="'guidance_' + id"
+			id="{{ id }}"
+			(blur)="checkErrors()"
+			(change)="checkErrors()"
+			#fudisTextInput
+			[type]="type"
+			[attr.required]="required"
+			[formControl]="fudisFormControl"
+			[ngClass]="classes"
+			class="fudis-text-input__input"
+			[attr.minlength]="minLength"
+			[attr.maxlength]="maxLength" />
 
-	<mat-error class="fudis-text-input__error" *ngIf="fudisFormControl.hasError(type) && fudisFormControl.touched">
-		toka näkyy
-		{{ errorsToShow.message }}
-	</mat-error>
-
-	<mat-error class="fudis-text-input__error" *ngIf="fudisFormControl.hasError('required') && fudisFormControl.touched">
-		kolmas näkyy
-		{{ errorsToShow.message }}
-	</mat-error> -->
-
-	<!-- <ng-container *ngFor="let error of errorsToShow | keyvalue">
-		<mat-error
-			class="fudis-text-input__error"
-			*ngIf="(fudisFormControl.hasError(error.key) || fudisFormControl.hasError(type)) && fudisFormControl.touched">
-			{{ error.value }}
-		</mat-error>
-	</ng-container> -->
-
-	<!-- <mat-error
-		class="fudis-text-input__error"
-		*ngIf="fudisFormControl.hasError('email') && !fudisFormControl.hasError('required')">
-		Please enter a valid email address
-	</mat-error>
-	<mat-error
-		class="fudis-text-input__error"
-		*ngIf="!fudisFormControl.hasError('email') && fudisFormControl.hasError('required')"
-		>Pakollinen kenttä</mat-error
-	> -->
-	<mat-hint class="fudis-text-input__help-text" *ngIf="helpText">{{ helpText }}</mat-hint>
-	<mat-hint class="fudis-text-input__character-limit-indicator" *ngIf="characterLimitIndicatorValue" align="end"
-		>{{ fudisTextInput.value.length }} / {{ characterLimitIndicatorValue }}</mat-hint
-	>
-</mat-form-field>
+		<div id="guidance_{{ id }}" class="fudis-text-input__guidance">
+			<div
+				aria-live="polite"
+				role="alert"
+				class="fudis-text-input__guidance__error"
+				*ngIf="fudisFormControl.invalid && fudisFormControl.touched">
+				{{ this.defaultError }}
+			</div>
+			<div class="fudis-text-input__guidance__help-text">{{ helpText }}</div>
+		</div>
+	</label>
+</div>
