+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/text-input/text-input.component.ts
@@ -1,27 +1,32 @@
-import { Attribute, Component, Input, OnInit, ViewChild, ElementRef } from '@angular/core';
+// eslint-disable-next-line max-classes-per-file
+import { Component, Input, Output, OnInit, EventEmitter, ViewChild, ElementRef } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
 
+type Error = {
+	id: string;
+	message: string;
+};
 @Component({
 	selector: 'fudis-text-input',
 	templateUrl: './text-input.component.html',
 	styleUrls: ['./text-input.component.scss'],
 })
 export class TextInputComponent implements OnInit {
-	constructor(@Attribute('required') required: boolean | '' | 'true') {
-		if (required === '' || required === 'true' || required) {
-			this.required = true;
-		} else {
-			this.required = false;
-		}
-	}
-
 	// Bind input field
 	@ViewChild('fudisTextInput') input: ElementRef<HTMLInputElement>;
+
+	@Output() errorOutput: EventEmitter<Error> = new EventEmitter<Error>();
 
 	/**
 	 *	Label is mandatory for every input
 	 */
 	@Input() label: string;
+
+	@Input() inputId?: string;
+
+	@Input() required?: boolean = false;
+
+	@Input() size?: 's' | 'm' | 'l' = 'l';
 
 	/**
 	 *	Helper or info text for the input, aligned underneath the input
@@ -36,18 +41,20 @@ export class TextInputComponent implements OnInit {
 
 	@Input() characterLimitIndicatorValue?: string;
 
+	@Input() labelHeight?: 'single' | 'double' | 'triple' | 'unset' = 'unset';
+
 	/**
 	 *	Type of the input
 	 */
 	@Input() type: 'email' | 'number' | 'password' | 'tel' | 'text' | 'url' = 'text';
-
-	required: boolean;
 
 	validatorArray: Array<any> = [];
 
 	fudisFormControl = new FormControl('', this.validatorArray);
 
 	defaultError: string;
+
+	id: string;
 
 	ngOnInit() {
 		if (this.required) {
@@ -66,7 +73,26 @@ export class TextInputComponent implements OnInit {
 			this.validatorArray.push(Validators.pattern('^[0-9]*$'));
 		}
 		this.fudisFormControl = new FormControl('', this.validatorArray);
+
+		// Setting id's and names
+		if (this.inputId) {
+			this.id = this.inputId;
+		} else {
+			this.id = this.randomId();
+		}
 	}
+
+	// Temporary randomiser for ids
+	randomId = () => {
+		const idFromLabel = this.label
+			.toLowerCase()
+			.replace(/[^a-z0-9]+/g, ' ')
+			.split(' ')
+			.slice(0, 3)
+			.join('-');
+		const randomId = `${idFromLabel}-${Math.random().toString(36).slice(2, 6)}`;
+		return randomId;
+	};
 
 	// Check and returns browsers native error message
 	checkErrors(): void {
@@ -92,11 +118,20 @@ export class TextInputComponent implements OnInit {
 			}
 			this.defaultError = inputElement.validationMessage;
 		}
+
+		// Emit error to parent
+		if (this.fudisFormControl.invalid) {
+			this.getErrorOutput(this.id, this.defaultError);
+		}
+	}
+
+	getErrorOutput(id: string, error: string) {
+		this.errorOutput.emit({ id, message: error });
 	}
 
 	public get classes(): string[] {
 		if (this.fudisFormControl.touched && this.fudisFormControl.invalid) {
-			return ['fudis-text-input--invalid'];
+			return ['fudis-text-input__input--invalid'];
 		}
 		return [];
 	}
