--- ngx-fudis/projects/ngx-fudis/src/lib/components/form/vanilla-text-input/vanilla-text-input.component.html
@@ -1,32 +1,0 @@
-<div class="fudis-form-element fudis-text-input fudis-text-input__{{ size }}">
-	<label class="fudis-text-input__label" for="{{ id }}">
-		<div #fudisTextInputLabel class="fudis-text-input__label__content">
-			<span class="fudis-text-input__label__text fudis-text-input__label__text__{{ labelHeight }}">{{ label }}</span
-			><span class="fudis-text-input__label__required" *ngIf="required">Pakollinen&nbsp;*</span>
-		</div>
-		<input
-			[attr.aria-describedby]="'guidance_' + id"
-			id="{{ id }}"
-			(blur)="checkErrors()"
-			(change)="checkErrors()"
-			#fudisTextInput
-			[type]="type"
-			[attr.required]="required"
-			[formControl]="fudisFormControl"
-			[ngClass]="classes"
-			class="fudis-text-input__input"
-			[attr.minlength]="minLength"
-			[attr.maxlength]="maxLength" />
-
-		<div id="guidance_{{ id }}" class="fudis-text-input__guidance">
-			<div
-				aria-live="polite"
-				role="alert"
-				class="fudis-text-input__guidance__error"
-				*ngIf="fudisFormControl.invalid && fudisFormControl.touched">
-				{{ this.defaultError }}
-			</div>
-			<div class="fudis-text-input__guidance__help-text">{{ helpText }}</div>
-		</div>
-	</label>
-</div>
