--- ngx-fudis/projects/ngx-fudis/src/lib/components/form/vanilla-text-input/vanilla-text-input.component.scss
@@ -1,123 +1,0 @@
-@use '../../../foundations/borders/mixins.scss' as borders;
-@use '../../../foundations/spacing/tokens.scss' as spacing;
-@use '../../../foundations/typography/mixins.scss' as typography;
-@use '../../../foundations/typography/tokens.scss' as typographyTokens;
-@use '../../../foundations/colors/tokens.scss' as colors;
-@use '../../../foundations/colors/mixins.scss' as colorMixins;
-@use '../../../foundations/breakpoints/mixins.scss' as breakpoints;
-@use '../../../foundations/focus/mixins.scss' as focus; 
-
-.fudis-text-input {
-	display: grid;
-	max-width: calc(23 * spacing.$fudis-spacing-s);
-
-
-	&__l {
-		width: 100%;
-	}
-
-	&__m {
-		width: calc(14 * spacing.$fudis-spacing-s);
-	}
-
-	&__s {
-		max-width: calc(10 * spacing.$fudis-spacing-s);
-	}
-
-	&__label {
-		&__content {
-			@include colorMixins.fudis-text-color('gray-dark');
-
-			position: relative;
-			padding-bottom: spacing.$fudis-spacing-xs;
-			overflow: hidden;
-		}
-
-		&__required {
-			@include typography.fudis-body-s-regular;
-
-			display: inline-block;
-			position: absolute;
-			right: 0;
-			bottom: 0;
-			text-transform: initial;
-			
-		}
-
-		&__text {
-			@include typography.fudis-text-field-label;
-
-			display: flex;
-			align-items: flex-end;
-			overflow: hidden;
-
-			&__unset {
-				height: initial;
-			}
-
-			&__single {
-				@include breakpoints.fudis-breakpoint('m') {
-					height: calc(typographyTokens.$fudis-text-input-line-height)
-				}
-			}
-
-			&__double {
-				@include breakpoints.fudis-breakpoint('m') {
-					height: calc(2 * typographyTokens.$fudis-text-input-line-height);
-				}
-
-			}
-
-			&__triple {
-				@include breakpoints.fudis-breakpoint('m') {
-					height: calc(3 * typographyTokens.$fudis-text-input-line-height);
-				}
-			}
-
-			
-		}
-	}
-
-	&__input {
-		@include borders.fudis-border('1px', 'solid', 'gray-middle');
-		@include typography.fudis-text-field-input;
-		@include borders.fudis-border-radius('2px');
-
-		box-sizing: border-box;
-		padding: spacing.$fudis-spacing-xxs spacing.$fudis-spacing-s;
-		width: 100%;
-		height: spacing.$fudis-spacing-xl;
-
-		&:focus {
-			@include focus.fudis-focus-form-field;
-		}
-
-		&--invalid {
-			border-color: colors.$fudis-red;
-			 }	
-	}
-
-	&__guidance {
-		&__help-text {
-			@include colorMixins.fudis-text-color('gray-dark');
-			@include typography.fudis-body-s-regular;
-	
-			margin-top: spacing.$fudis-spacing-xs;
-		}
-		
-		&__character-limit-indicator {
-			@include colorMixins.fudis-text-color('gray-dark');
-		}
-	
-		&__error {
-			@include typography.fudis-body-s-regular;
-			@include colorMixins.fudis-text-color('red');
-	
-			margin-top: spacing.$fudis-spacing-xs;
-		}
-
-	}
-
-	
-}
-
