--- ngx-fudis/projects/ngx-fudis/src/lib/components/form/vanilla-text-input/vanilla-text-input.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { VanillaTextInputComponent } from './vanilla-text-input.component';
-
-describe('VanillaTextInputComponent', () => {
-	let component: VanillaTextInputComponent;
-	let fixture: ComponentFixture<VanillaTextInputComponent>;
-
-	beforeEach(async () => {
-		await TestBed.configureTestingModule({
-			declarations: [VanillaTextInputComponent],
-		}).compileComponents();
-	});
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(VanillaTextInputComponent);
-		component = fixture.componentInstance;
-		component.label = 'Label for testing purposes';
-		fixture.detectChanges();
-	});
-
-	it('should create', () => {
-		expect(component).toBeTruthy();
-	});
-});
