+++ ngx-fudis/projects/ngx-fudis/src/lib/components/icon/icon.component.ts
@@ -1,99 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-export type Icon =
-	| 'achievement'
-	| 'achievement-small'
-	| 'alert'
-	| 'alert-fill'
-	| 'alert-fill-attention'
-	| 'alert-small'
-	| 'archive'
-	| 'arrow-big'
-	| 'arrow-dashed'
-	| 'arrow-solid'
-	| 'back'
-	| 'brochure'
-	| 'bullets'
-	| 'calendar'
-	| 'change-log'
-	| 'check'
-	| 'check-indeterminate-small'
-	| 'check-small'
-	| 'checkmark-circle'
-	| 'chevron'
-	| 'chevron-ring'
-	| 'chevron-ring-fill'
-	| 'clock'
-	| 'close'
-	| 'close-big'
-	| 'code'
-	| 'decrease-indent'
-	| 'delete'
-	| 'dot'
-	| 'edit'
-	| 'editor'
-	| 'exclamation-mark-circle'
-	| 'exclamation-mark-circle-fill'
-	| 'exclamation-mark-small'
-	| 'external'
-	| 'eye'
-	| 'eye-blind'
-	| 'fail'
-	| 'hourglass'
-	| 'increase-indent'
-	| 'info'
-	| 'info-circle'
-	| 'info-circle-fill'
-	| 'info-small'
-	| 'junction'
-	| 'link'
-	| 'list-add'
-	| 'list-minus'
-	| 'lock'
-	| 'lock-open'
-	| 'magic-wand'
-	| 'mail'
-	| 'menu'
-	| 'message'
-	| 'minus'
-	| 'minus-ring-fill'
-	| 'notebook'
-	| 'notification'
-	| 'numbering'
-	| 'pdf'
-	| 'people'
-	| 'person'
-	| 'person-small'
-	| 'picker'
-	| 'pin-small'
-	| 'place'
-	| 'place-ring-fill'
-	| 'plus'
-	| 'print'
-	| 'question-mark'
-	| 'question-mark-small'
-	| 'required'
-	| 'ring-close'
-	| 'ring-close-fill'
-	| 'ring-plus'
-	| 'ring-plus-fill'
-	| 'rosette'
-	| 'rule'
-	| 'search'
-	| 'seats'
-	| 'settings'
-	| 'shopping-cart'
-	| 'sorter'
-	| 'star'
-	| 'switch'
-	| 'three-dots'
-	| 'three-dots-small'
-	| 'waiting-approval'
-	| 'waiting-decline'
-	| 'zoom-in'
-	| 'zoom-out';
-
-export type IconColor = 'attention' | 'danger' | 'default' | 'light' | 'primary' | 'success' | 'white';
+import { Icon, IconColor } from '../../types/icons';
 
 /**
  * Example: `<fudis-icon icon="info" color="primary"></fudis-icon>`
