+++ ngx-fudis/projects/ngx-fudis/src/lib/components/typography/body-text/body-text.stories.ts
@@ -14,7 +14,7 @@ const Template: Story<BodyTextComponent> = () => ({
 `,
 });
 
-export const BodyTextExample = Template.bind({});
-BodyTextExample.args = {
+export const BodyText = Template.bind({});
+BodyText.args = {
 	size: 'l-regular',
 };
