+++ ngx-fudis/projects/ngx-fudis/src/lib/components/typography/text-spacing/text-spacing.stories.ts
@@ -42,5 +42,5 @@ export const Default = () => ({
 });
 
 Default.story = {
-	name: 'Example',
+	name: 'Text Spacing',
 };
