+++ ngx-fudis/projects/ngx-fudis/src/lib/foundations/borders/readme.stories.mdx
@@ -4,7 +4,7 @@ import { Meta } from '@storybook/addon-docs';
 
 # Border mixins
 
-When using border mixins load them with `@use '../../../foundations/borders/mixins.scss' as borders;` to the component's .scss file.
+When using border mixins load them with `@use '../path/to/foundations/borders/mixins.scss' as borders;` to the component's .scss file.
 
 ## Border
 Set the border of an element.
