+++ ngx-fudis/projects/ngx-fudis/src/lib/foundations/breakpoints/readme.stories.mdx
@@ -4,7 +4,7 @@ import { Meta } from '@storybook/addon-docs';
 
 # Breakpoint mixins
 
-When using breakpoint mixins load them with `@use '../../../foundations/breakpoints/mixins.scss' as breakpoints;` to the component's .scss file.
+When using breakpoint mixins load them with `@use '../path/to/foundations/breakpoints/mixins.scss' as breakpoints;` to the component's .scss file.
 
 The mixins are using only min-width, so when you are using the mixin **fudis-breakpoint('m')**, your styles will be applied from the medium breakpoint and up.
