--- ngx-fudis/projects/ngx-fudis/src/lib/foundations/shadows/readme.stories.mdx
@@ -1,7 +1,0 @@
-import { Meta } from '@storybook/addon-docs';
-
-<Meta title="Foundations/Shadows" />
-
-# HELLO FROM SHADOWS FOUNDATIONS!
-
-Bacon ipsum dolor amet pancetta strip steak picanha doner, biltong tenderloin turducken turkey porchetta burgdoggen pastrami cow ribeye. Swine kielbasa hamburger, shoulder drumstick beef ribs strip steak flank t-bone andouille prosciutto turducken meatloaf fatback ball tip. Leberkas rump chicken, shoulder hamburger chuck salami pork chop porchetta biltong. Jerky pork chop leberkas tail drumstick, shoulder hamburger strip steak. Prosciutto jowl bacon biltong cupim pork. Turkey strip steak shankle leberkas chicken, filet mignon cow andouille meatloaf t-bone turducken burgdoggen.
