+++ ngx-fudis/projects/ngx-fudis/src/lib/foundations/typography/typography.stories.mdx
@@ -1,42 +1,40 @@
 import { Meta, Typeset } from '@storybook/addon-docs/blocks';
 
-
-<Meta title="Foundations/Typography"/>
+<Meta title="Foundations/Typography" />
 
 export const heading = {
-  type: {
-    primary: "Fira Sans",
-  },
-  weight: {
+	type: {
+		primary: 'Fira Sans',
+	},
+	weight: {
 		medium: '500',
-    bold: '600',
-  },
-  size: {
-    xxs: 14,
-    xs: 16,
-    s: 18,
-    m: 20,
-    l: 24,
-    xl: 28,
-    xxl: 36,
-  },
+		bold: '600',
+	},
+	size: {
+		xxs: 14,
+		xs: 16,
+		s: 18,
+		m: 20,
+		l: 24,
+		xl: 28,
+		xxl: 36,
+	},
 };
 
 export const body = {
-  type: {
-    primary: "Fira Sans",
-  },
-  weight: {
+	type: {
+		primary: 'Fira Sans',
+	},
+	weight: {
 		light: '300',
-    regular: '400',
-  },
-  size: {
+		regular: '400',
+	},
+	size: {
 		l: 16,
 		m: 14,
 		s: 12,
-  },
+	},
 };
-
 
 export const text_xxl = 'Fudis heading XXL (2.25rem) 500';
 export const text_xl = 'Fudis heading XL (1.75rem) 500';
@@ -46,12 +44,12 @@ export const text_s = 'Fudis heading S (1.125rem) 500';
 export const text_xs = 'Fudis heading XS (1rem) 600';
 export const text_xxs = 'Fudis heading XXS (0.857rem) 600';
 
-export const body_s = 'Fudis body s-regular (0.75rem) 400: "The quick brown fox jumps over the lazy dog."'
-export const body_s_light = 'Fudis body s-light (0.75rem) 300: "The quick brown fox jumps over the lazy dog."'
-export const body_m = 'Fudis body m-regular (0.875rem) 400: "The quick brown fox jumps over the lazy dog."'
-export const body_m_light = 'Fudis body m-light (0.875rem) 300: "The quick brown fox jumps over the lazy dog."'
-export const body_l = 'Fudis body l-regular (1rem) 400: "The quick brown fox jumps over the lazy dog."'
-export const body_l_light = 'Fudis body l-light (1rem) 300: "The quick brown fox jumps over the lazy dog."'
+export const body_s = 'Fudis body s-regular (0.75rem) 400: "The quick brown fox jumps over the lazy dog."';
+export const body_s_light = 'Fudis body s-light (0.75rem) 300: "The quick brown fox jumps over the lazy dog."';
+export const body_m = 'Fudis body m-regular (0.875rem) 400: "The quick brown fox jumps over the lazy dog."';
+export const body_m_light = 'Fudis body m-light (0.875rem) 300: "The quick brown fox jumps over the lazy dog."';
+export const body_l = 'Fudis body l-regular (1rem) 400: "The quick brown fox jumps over the lazy dog."';
+export const body_l_light = 'Fudis body l-light (1rem) 300: "The quick brown fox jumps over the lazy dog."';
 
 # Typography
 
@@ -62,124 +60,96 @@ export const body_l_light = 'Fudis body l-light (1rem) 300: "The quick brown fox
 **Weights:** 500(medium), 600(bold)
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.xxl),
-  ]}
-  fontWeight={heading.weight.medium}
-  sampleText={text_xxl}
-  fontFamily={heading.type.primary}
-/>
-
-<Typeset
-  fontSizes={[
-    Number(heading.size.xl),
-  ]}
-  fontWeight={heading.weight.medium}
-  sampleText={text_xl}
+	fontSizes={[Number(heading.size.xxl)]}
+	fontWeight={heading.weight.medium}
+	sampleText={text_xxl}
 	fontFamily={heading.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.l),
-  ]}
-  fontWeight={heading.weight.medium}
-  sampleText={text_l}
+	fontSizes={[Number(heading.size.xl)]}
+	fontWeight={heading.weight.medium}
+	sampleText={text_xl}
 	fontFamily={heading.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.m),
-  ]}
-  fontWeight={heading.weight.medium}
-  sampleText={text_m}
+	fontSizes={[Number(heading.size.l)]}
+	fontWeight={heading.weight.medium}
+	sampleText={text_l}
 	fontFamily={heading.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.s),
-  ]}
-  fontWeight={heading.weight.medium}
-  sampleText={text_s}
+	fontSizes={[Number(heading.size.m)]}
+	fontWeight={heading.weight.medium}
+	sampleText={text_m}
 	fontFamily={heading.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.xs),
-  ]}
-  fontWeight={heading.weight.bold}
-  sampleText={text_xs}
+	fontSizes={[Number(heading.size.s)]}
+	fontWeight={heading.weight.medium}
+	sampleText={text_s}
 	fontFamily={heading.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(heading.size.xxs),
-  ]}
-  fontWeight={heading.weight.bold}
-  sampleText={text_xxs}
+	fontSizes={[Number(heading.size.xs)]}
+	fontWeight={heading.weight.bold}
+	sampleText={text_xs}
 	fontFamily={heading.type.primary}
 />
 
+<Typeset
+	fontSizes={[Number(heading.size.xxs)]}
+	fontWeight={heading.weight.bold}
+	sampleText={text_xxs}
+	fontFamily={heading.type.primary}
+/>
 
 ## Body Text / Leipäteksti
 
 **Weights:** 300(light) 400(regular)
 
-
 <Typeset
-  fontSizes={[
-    Number(body.size.l),
-  ]}
-  fontWeight={body.weight.regular}
-  sampleText={body_l}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.l)]}
+	fontWeight={body.weight.regular}
+	sampleText={body_l}
+	fontFamily={body.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(body.size.l),
-  ]}
-  fontWeight={body.weight.light}
-  sampleText={body_l_light}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.l)]}
+	fontWeight={body.weight.light}
+	sampleText={body_l_light}
+	fontFamily={body.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(body.size.m),
-  ]}
-  fontWeight={body.weight.regular}
-  sampleText={body_m}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.m)]}
+	fontWeight={body.weight.regular}
+	sampleText={body_m}
+	fontFamily={body.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(body.size.m),
-  ]}
-  fontWeight={body.weight.light}
-  sampleText={body_m_light}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.m)]}
+	fontWeight={body.weight.light}
+	sampleText={body_m_light}
+	fontFamily={body.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(body.size.s),
-  ]}
-  fontWeight={body.weight.regular}
-  sampleText={body_s}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.s)]}
+	fontWeight={body.weight.regular}
+	sampleText={body_s}
+	fontFamily={body.type.primary}
 />
 
 <Typeset
-  fontSizes={[
-    Number(body.size.s),
-  ]}
-  fontWeight={body.weight.light}
-  sampleText={body_s_light}
-  fontFamily={body.type.primary}
+	fontSizes={[Number(body.size.s)]}
+	fontWeight={body.weight.light}
+	sampleText={body_s_light}
+	fontFamily={body.type.primary}
 />
